login and registration controllers moved to seoarate files
