+++ controllers/users.js
@@ -1,42 +1,6 @@
-const jwt = require('jsonwebtoken');
-const bcrypt = require('bcryptjs');
 const User = require('../models/user');
 require('dotenv').config();
 const BadRequestError = require('../middlewares/BadRequestError');
-const UnauthorizedError = require('../middlewares/UnauthorizedError');
-const ConflictError = require('../middlewares/ConflictError');
-
-const { NODE_ENV, JWT_SECRET } = process.env;
-
-module.exports.createUser = async (req, res, next) => {
-  try {
-    const {
-      name, email, password,
-    } = req.body;
-    const emailValidation = await User.findOne({ email });
-    if (emailValidation) {
-      next(new ConflictError('Такой пользователь уже существует'));
-      return;
-    }
-    const hash = await bcrypt.hash(password, 10);
-    const user = await User.create({
-      name,
-      email,
-      password: hash,
-    });
-    const newUser = {
-      name: user.name,
-      email: user.email,
-      _id: user._id,
-    };
-    res.status(201).send({ data: newUser });
-  } catch (err) {
-    if (err.name === 'ValidationError') {
-      next(new BadRequestError('Переданы некорректные данные'));
-    }
-    next(err);
-  }
-};
 
 module.exports.updateUser = async (req, res, next) => {
   console.log(req.user._id);
@@ -53,16 +17,6 @@ module.exports.updateUser = async (req, res, next) => {
       next(new BadRequestError('Переданы некорректные данные'));
     }
     next(err);
-  }
-};
-
-module.exports.login = async (req, res, next) => {
-  const { email, password } = req.body;
-  try {
-    const user = await User.findUserByCredentials(email, password);
-    res.send({ token: jwt.sign({ _id: user._id }, NODE_ENV === 'production' ? JWT_SECRET : 'dev-secret', { expiresIn: '7d' }) });
-  } catch (err) {
-    next(new UnauthorizedError('Неверный логин или пароль'));
   }
 };
